Move analytic model files and update README
This commit moves all files for the analytic model into a new directory
ahead of the merge with the agent based model. Minor changes to a couple
of files are included. The analytic folder now has the original README,
while a new, stripped down README has been created in the base directory.