Massive amount of bad-math and doc corrections
The litany of things that were wrong with how this model ran before this
commit are long and arduous. Integration against time is now correct and
adjusts automatically for different sample timestepping. Spatial
integration is performed in fortran via the scipy.stats mvn package (which
is hidden, but seems to be due to Dr. Alan Genz and rather accurate), so
that the key function now returns a probability mass function over the
cells, sensitive to spatial scaling. Advection has been fixed - it seems
that before, I was completely ignoring the fact that the units for wind
would be km/hr, which would need to be converted to m/unit time. Functions
have been added, deleted, and renamed to reflect all of these changes. The
spatial integration is going to end up being the slow part, but I cannot
pass vectors into mvn so the loop is unavoidable. numba can't jit compile
it in nopython mode because it doesn't know mvn. But speeding up the code
can safely be a todo, as long as the result is correct.