Bug fixes, adds precision sensitive integration
Previously, the code was not converting from (x,y) coordinates (in meters
from release point) to row,column comprehension in ndarrays. This should
be fixed, but needs more testing. Adds get_mvn_cdf_values which takes care
of precision sensitive integration of the bivariate normal cdf centered at
the origin. This is MUCH faster than the previous way of doing things.
TODO is to not round advection to the nearest cell center, and instead
pass the remainder to get_mvn_cdf_values as a mean.