Adds get_wind_data, read_wind_file days output,etc
This commit adds the function get_wind_data, but it is not yet integrated
into the rest of the model or tested. It's purpose is to interpolate the
wind data for more take-off time possibilities and a smoother model
output, (hopefully) making it easier to fit the model parameters to data.
read_wind_file now outputs a sorted list of the days included (which are
the keys to the wind_data dict). Tests and plotting has been updated
accordingly, and some housekeeping done in the testing suite.