Update h to scale w/ f, incorporate get_wind_data
Interpolation using get_wind_data is now incorporated into the code,
including in the testing suite and the plotting suite. On introspection of
the results, it became apparent that extra flight probability could build
up toward the end of the day in h_flight_prob, resulting in a considerable
chance of flights in the middle of the night. This has been fixed in both
the ipython notebook and the model code by scaling the added probability,
integral_avg, f scaled by f_max. All tests now pass and plotting the
result yields in the expected behavior. This commit also fixes a bug by
which Plot_ParasitoidModel.plot_prob_mass() plotted upside-down.