Adds Run.py to run model, save result
Run.py includes a class file to keep track of parameter values. These
values can be set either by editing the source or by passing changes from
the default in through command line arguments (not yet fully implemented).
the Params class includes an option to run fewer than the full number of
days in the wind data. Results are saved in a numpy .npz file, along with
the list of parameters in a json file. In a future commit, it would be
nice to be able to import a json file to the Params class as another way
of loading parameters for a simulation run.