Major restruct, adds CalcSol. fft sol persistance
This commit includes two major changes.

1. Restructure the code so that methods concerning the convolution part of
the simulation are in their own module, CalcSol. This allows us to
separate all the fft and cuda stuff from the model definitions.

2. The convolution is now by fft and ifft, and the fft solution of each
day is held so that it can be updated with the next day's filter. The idea
here is that we should only be doing one fft and one ifft for each day.