Moves all fft to CalcSol, r_small_vals much faster
Moves the entire convolution cycle, including fft,fft convolution, ifft,
and reduction of solutions to a serial process which takes place in
CalcSol.py. The speed gain from multiprocessing these steps was
negligable, and it is expected that doing all of these steps on the GPU
at once (especially including the persistance of the current fft solution
on the GPU) will result in the fastest simulation speeds. Solution
reduction has shown itself to be necessary in this cycle due to RAM
limitations, and a this commit greatly increases the speed of this process
(by as much as five times).