Adds file_read_chg to Params cls, plot prv results
Previously saved results (as an npz and a json parameter file) can now be
plotted by calling running Plot_Result.py and passing in the file path as
an argument. Most of the functionality for this is contained in the new
Plot_Result.py functions plot and main, though the routines also call upon
the Params class from Run.py to load the saved paramters using the new
file_read_chg method.