Adds Reikna CUDA, passes fft,conv tests
scikit-cuda does not have a 2D fft routine, and pyfft has been
discontinued, absorbed into Reikna, which is a library with
various GPU algorithms built on top of PyCUDA and PyOpenCL. Reikna
is now the engine for the CUDA routines in the model, and as of
this commit, it passes tests for correct 2d fft convolution on the
GPU. Run.py is broken on the GPU as of this commit and needs to be
debugged. Other changes reflect a move to linux that may need to be
addressed for windows compatability.