Updates Plot_ParasitoidModel API, check bnds in pmf
Parameter values for Plot_ParasitoidModel are now contained within a
parameter object defined at the top of the module. This object is
automatically created when the module is imported so that all functions in
the module can be run without explicitly passing any parameters. This
paradigm allows parameters to be easily modified on the fly in ipython for
experimentation, and also for multiple parameter sets to be formed and
compared (via multiple objects).