Fixes massive slowdown on nemo
Somehow, passing a specific number of processors to Pool when setting up
a worker pool was massively slowing down the first phase of the model
on nemo. Some waiting worker was involved, though this still does not
necessarily explain the massive slowdowns before the multiprocessing
commit, though this could have been solved by a restart. Anyway, the
issue seems to be fixed in this commit.