back_solve test off GPU - bug fixes
Implements test_back_solve in test_CalcSol.py to test the non-GPU
back_solve routine with sparse domain signals/filters. Debugs
CalcSol.back_solve() so that it is passing the tests. Some things that
still need to be done: GPU test, routine to deal with periodic boundary
conditions in the convolution (since the fft of the convolution is
maintained between multiplications, the bndry condition could be an issue
when the signal gets close to the bndry), GPU solutions are not currently
reduced for negligible values.