Adds get_fields, major restructure of Bayes_MCMC
For now, Bayes_MCMC will be a script instead of a file to be run... the
examples in the PyMC book are all scripts, and it was becoming a chore to
try and organize everything with a main() function. Once everything is
written, a main() wrapper can easily be included.