Closes #8, diffusion outside of flow
For now, the diffusion coefficients outside of flow are assumed to be the
same as within flow. Plotting now ignores a 9x9 square about the origin
when finding the maximum scale value to accomodate the action at the
origin while still revealing structure elsewhere. Since the total origin
population/probability is reduced, the mask value has also been raised in
comparsion to the max value in the domain to three orders of magnitude
below the max. If this is larger than 1, the mask value is 1 instead.