Improves error reporting in Params.cmd_line_chg()
Apparently, command line arguments in Windows are broken up not just over
spaces, but also over parenthesis and commas. This makes passing a tuple
as a command line argument difficult. Since the unix behavior is unknown,
the best thing for the moment is just to return an informative error
message when this happens: enclosing the offendign argument in quotations
solves the problem.