Major CUDA bug fix and closes #1 (testing needed)
Prior to this commit, back_solve in cuda_lib was overwriting the
persistant solution instead of making a copy and operating on that to get
each backsolve. This will have resulted in very incorrect results for CUDA
based population simulations. This was discovered and fixed while
addressing issue #1, which this commit closes. It has not yet been tested
on a GPU, but it has been tested off GPU (CalcSol.py only) and performs
fine there as evidenced by running simulations on small enough domains
where the boundary conditions clearly come into play.