Reverts .copy() in cuda_lib
Apparently that major fix was completely unnecessary, and for some unknown
reason .copy in PyCUDA completely breaks the result. In this version, CUDA
and non-CUDA results are a close approximation to each other... CUDA
returns a somewhat deflated value at the origin in the population model,
but it may be attributed to round-off error in 32-bit floats.