Major Bayes bug fix, kalbar release fix, begins adding grid sampling
Many things happening at once here, since one thing led to the discovery
of another:
1. popdensity_to_emergence now assumes that emergence collections are
done at the beginning of the day and emergence observations at the end.
This reflects data, and required a re-write of the emergence projection
algorithm, which was already buggy for other reasons.
2. kalbar wind data did not start until midnight after the release. As a
result, we will assume that the release was actually at the time the wind
data started - midnight of the next day. We will further assume that the
release was only conducted over that one day. Hopefully these assumptions
will average out the reality of the situation.
3. This commit adds data on adult wasps sampling and adds pandas code to
LocInfo under a method. This method isn't called yet, and is currently
broken.
4. Tests were updated as appropriate and pass. A comment was added in
ParasitoidModel.py.