Critical bug fix in population model - negative pop
Previous to this fix, r_small_vals in CalcSol was assuming that population
arrays were probabilility mass functions and thus evenly distributing 1
minus the sum of the array across the array. This resulted in negative
population numbers for much of the solution. All population model results
prior to this bug fix should be considered invalid.