Bayes_Run: n_periods for mu_r, drops cardinal data, with Pool
Switches out the n_periods variable for the mu_r variable as a more
interpretable value. In a later commit, we will use a coarser (10m**2)
grid in Run.py to speed things up since the MCMC stuff is going
unacceptably slow, so we are dropping the cardinal data as part of the
parameter fitting at this time. At 10m**2, there just isn't enough there
to bother messing with. Also, the multiprocessing pool is now used in a
context manager for more readable code and better exception handling.
Finally, some priors have been changed, including quite a bump in
n_periods to 30 min/flight. This is in response to our model not reaching
the furthest field at all, especially not with any population numbers that
would not be dropped by float32 roundoff.