n_periods -> 30, coarser domain, cdf_eps to 0.001
After some experimentation, n_periods has been set to 30 so that the model
will reach all sentinel fields by the end of the model run. The domain has
been made coarser to mitigate some of the issues with float32 variables in
the GPU solver giving significantly different results at small numbers
when compared to its float64 counterpart. This will also be rather useful
for speed when trying some of the larger spread kernels in the Bayesian
algorithms - which is also the thinking behind changing cdf_eps from
0.0001 to 0.001 in ParasitoidModel.get_mvn_cdf_values. The difference
seems to be rather minor when comparing sentinel field counts before and
after the change - typically the population in a sentinel field differs by
-3 orders of magnitude from the total.