Establishes one Pool for the entire module
This in an attempt to prevent the random sleeps that seem to be occurring
on nemo. Since these sleeps do not time out, it may be that rather than
occurring in the wait for jobs to finish, their occurance is in some way
related to constantly creating/breaking down the pool of workers. This
causes overhead anyway, so we should try not to do this - using one pool
of workers that persists throughout.