Drops sentinel field 'G' from data import
Due to certain traces not converging in the MCMC algorithm when most
others are, we suspect that non-linear fluid flow above the landscape may
have played a key role in spreading parasitoids to the furthest sentinel
field. This commit removes this field from the data (within Data_Import)
so that we can fit the remaining fields to our model.