'G' field included, Beta prior on correlations
The 'G' field is reincluded in Data_Import, since its exclusion does not
seem to be having the desired effect in the MCMC. After successfully
running MAP for the first time, it is clear that some sort of prior is
needed on the diffusion correlation values, since these are being pushed
to -1 and 1. This commit uses a transformed Beta prior with parameters
alpha=5 and beta=5.