Adds even stricter prior to mu_l_r
It's important that only the field level data is affecting this value, and
a strict prior may be our only shot of making this happen. We need to
enforce that this not another landscape-level mode of spread to be added
onto the one we already have, but something that is intrinsically linked
to the local scale.