Removes obs prob truncation, adds restricted priors
The truncation of the Beta distributions for the observation probabilities
was quite artificial - this commit removes that mechanic and ratchets up
the beta parameter to utilize a more informative prior instead. Starting
values have also been reduced by an order of magnitude. For the release
field, the Beta distributions are unchanged, but the starting values were
chosen to represent what we have been getting from the data... so
hopefully the finer epsilon value will not affect performance.