Changes Bayes_Run to use AdaptiveMetropolis
The reason MCMC was calculating around 12 model realizations per iteration
is that the Metropolis algorithm updates and checks each parameter step
invidually by design. The adaptive Metropolis method is the same method
except parameters are block updated from a joint normal distribution,
which in turn requires that a covariance matrix be specified and refined
as the algorithm continues. All stochastic variables informing the
deterministic spread model are now block updated in this way, which should
speed up the MCMC algorithm by an order of magnitude.