Surface plot adjustments, Plot_Result error handling
Patches on the surface plot (matching the bar underneath) are now opaque.
View angle for the 3D plot is changed, as are the margins of the figure to
accomodate this.
Plot_Result now has error handling to deal with an entered date that is
out of the range of the simulation.